refactor(points): unify orderNumber type and adjust model & UI
- Change `PointsModel` default `orderNumber` from 0.0 to use `qrCodeId` (int)
- Convert `PointsEntity.orderNumber` type from `double` to `int`
- Update sample `_points` data in `PointsScreen` to use integer orderNumber values
- Import and use new `AppColor.grey` for orderNumber text in `PointsWidget`
- Add `grey` constant to `AppColor`
- Correct indentation in generated `l10n.dart` around `loyalty_program_for_PetroJac…` getter

+++ lib/features/dashboard/domain/entity/points_entity.dart
@@ -5,7 +5,7 @@ import '../../../../core/utils/types/points_status_enum.dart';
 class PointsEntity extends Equatable {
   final PointsStatusEnum orderStatus;
   final int points;
-  final double orderNumber;
+  final int orderNumber;
   final DateTime date;
 
   const PointsEntity({
